+++ packages/graphql-toolbox/src/modules/Login/Login.tsx
@@ -17,8 +17,7 @@
  * limitations under the License.
  */
 
-import { useCallback } from "react";
-import { useContext, useState } from "react";
+import { useCallback, useContext, useState } from "react";
 import { FormInput } from "./FormInput";
 import { Button, HeroIcon } from "@neo4j-ndl/react";
 import { DEFAULT_BOLT_URL, DEFAULT_USERNAME } from "../../constants";
@@ -54,7 +53,7 @@ export const Login = () => {
                     url,
                 });
             } catch (error) {
-                setError((error as Error).message as string);
+                setError((error as Error).message);
             } finally {
                 setLoading(false);
             }
@@ -147,6 +146,7 @@ export const Login = () => {
                                                 className="underline"
                                                 href="https://neo4j.com/developer/javascript/#driver-configuration"
                                                 target="_blank"
+                                                rel="noreferrer"
                                             >
                                                 here
                                             </a>