+++ packages/graphql/src/graphql/directives/callback.ts
@@ -20,10 +20,13 @@
 import { DirectiveLocation, GraphQLDirective, GraphQLList, GraphQLNonNull, GraphQLString } from "graphql";
 import { CallbackOperationEnum } from "./arguments/enums/CallbackOperation";
 
+const description = "NOTE: this directive has been deprecated and will be removed in @neo4j/graphql version 4.0." +
+    "Instructs @neo4j/graphql to invoke the specified callback function when updating or creating the properties on a node or relationship."
+
+/** Deprecated in favour of @populatedBy */
 export const callbackDirective = new GraphQLDirective({
     name: "callback",
-    description:
-        "Instructs @neo4j/graphql to invoke the specified callback function when updating or creating the properties on a node or relationship.",
+    description,
     locations: [DirectiveLocation.FIELD_DEFINITION],
     args: {
         operations: {