+++ packages/graphql/src/translate/connection-clause/create-connection-clause.ts
@@ -17,6 +17,7 @@
  * limitations under the License.
  */
 
+import type { Integer } from "neo4j-driver";
 import type { ResolveTree } from "graphql-parse-resolve-info";
 import type { ConnectionField, ConnectionWhereArg, Context } from "../../types";
 import type { Node } from "../../classes";
@@ -25,7 +26,7 @@ import * as CypherBuilder from "../cypher-builder/CypherBuilder";
 import { hasExplicitNodeInInterfaceWhere } from "../where/property-operations/create-connection-operation";
 import { getOrCreateCypherNode } from "../utils/get-or-create-cypher-variable";
 import { createSortAndLimitProjection } from "./create-sort-and-limit";
-// eslint-disable-next-line import/no-cycle
+
 import { createEdgeSubquery } from "./create-edge-subquery";
 
 export function createConnectionClause({
@@ -33,11 +34,13 @@ export function createConnectionClause({
     field,
     context,
     nodeVariable,
+    returnVariable,
 }: {
     resolveTree: ResolveTree;
     field: ConnectionField;
     context: Context;
     nodeVariable: string;
+    returnVariable: CypherBuilder.Variable;
 }): CypherBuilder.Clause {
     if (field.relationship.union || field.relationship.interface) {
         return createConnectionClauseForUnions({
@@ -45,10 +48,12 @@ export function createConnectionClause({
             field,
             context,
             nodeVariable,
+            returnVariable,
         });
     }
 
     const whereInput = resolveTree.args.where as ConnectionWhereArg;
+    const firstArg = resolveTree.args.first as Integer | number | undefined;
     const relatedNode = context.nodes.find((x) => x.name === field.relationship.typeMeta.name) as Node;
     const edgeItem = new CypherBuilder.NamedVariable("edge");
 
@@ -70,12 +75,11 @@ export function createConnectionClause({
     ]);
 
     const totalCount = new CypherBuilder.NamedVariable("totalCount");
-
     const withSortAfterUnwindClause = createSortAndLimitProjection({
         resolveTree,
         relationshipRef: edgeItem,
         nodeRef: edgeItem.property("node"),
-        limit: relatedNode?.queryOptions?.getLimit(),
+        limit: relatedNode?.queryOptions?.getLimit(firstArg), // `first` specified on connection field in query needs to be compared with existing `@queryOptions`-imposed limit
         extraFields: [totalCountItem],
     });
 
@@ -93,7 +97,7 @@ export function createConnectionClause({
             edges: edgesList,
             totalCount,
         }),
-        resolveTree.alias,
+        returnVariable,
     ]);
     return CypherBuilder.concat(edgeSubquery, withClause, unwindSortClause, returnClause);
 }
@@ -103,11 +107,13 @@ function createConnectionClauseForUnions({
     field,
     context,
     nodeVariable,
+    returnVariable,
 }: {
     resolveTree: ResolveTree;
     field: ConnectionField;
     context: Context;
     nodeVariable: string;
+    returnVariable: CypherBuilder.Variable;
 }) {
     const whereInput = resolveTree.args.where as ConnectionWhereArg;
     const relatedNode = context.nodes.find((x) => x.name === field.relationship.typeMeta.name) as Node;
@@ -152,7 +158,6 @@ function createConnectionClauseForUnions({
 
     let withOrderClause: CypherBuilder.Clause | undefined;
     const limit = relatedNode?.queryOptions?.getLimit();
-
     const withOrder = createSortAndLimitProjection({
         resolveTree,
         relationshipRef: edgeItem,
@@ -172,7 +177,7 @@ function createConnectionClauseForUnions({
             edges: edgesList,
             totalCount,
         }),
-        resolveTree.alias,
+        returnVariable,
     ]);
 
     return CypherBuilder.concat(unionClauses, withEdgesAndTotalCount, withOrderClause, returnClause);