+++ packages/graphql/tests/tck/connections/connect-or-create/connect-or-create.test.ts
@@ -93,12 +93,12 @@ describe("Create or Connect", () => {
                 WITH this0
                 CALL {
                     WITH this0
-                    MERGE (this0_movies_connectOrCreate_this0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
+                    MERGE (this0_movies_connectOrCreate0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this0.title = $this0_movies_connectOrCreate_param1
-                    MERGE (this0)-[this0_movies_connectOrCreate_this1:ACTED_IN]->(this0_movies_connectOrCreate_this0)
+                        this0_movies_connectOrCreate0.title = $this0_movies_connectOrCreate_param1
+                    MERGE (this0)-[this0_movies_connectOrCreate_this0:ACTED_IN]->(this0_movies_connectOrCreate0)
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this1.screentime = $this0_movies_connectOrCreate_param2
+                        this0_movies_connectOrCreate_this0.screentime = $this0_movies_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN this0
@@ -155,12 +155,12 @@ describe("Create or Connect", () => {
                 WITH this
                 CALL {
                     WITH this
-                    MERGE (this_movies0_connectOrCreate_this0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
+                    MERGE (this_movies0_connectOrCreate0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this0.title = $this_movies0_connectOrCreate_param1
-                    MERGE (this)-[this_movies0_connectOrCreate_this1:ACTED_IN]->(this_movies0_connectOrCreate_this0)
+                        this_movies0_connectOrCreate0.title = $this_movies0_connectOrCreate_param1
+                    MERGE (this)-[this_movies0_connectOrCreate_this0:ACTED_IN]->(this_movies0_connectOrCreate0)
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this1.screentime = $this_movies0_connectOrCreate_param2
+                        this_movies0_connectOrCreate_this0.screentime = $this_movies0_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN collect(DISTINCT this { .name }) AS data"
@@ -252,14 +252,14 @@ describe("Create or Connect", () => {
                 WITH this0
                 CALL {
                     WITH this0
-                    MERGE (this0_movies_connectOrCreate_this0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
+                    MERGE (this0_movies_connectOrCreate0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this0.createdAt = datetime(),
-                        this0_movies_connectOrCreate_this0.id = randomUUID(),
-                        this0_movies_connectOrCreate_this0.title = $this0_movies_connectOrCreate_param1
-                    MERGE (this0)-[this0_movies_connectOrCreate_this1:ACTED_IN]->(this0_movies_connectOrCreate_this0)
+                        this0_movies_connectOrCreate0.createdAt = datetime(),
+                        this0_movies_connectOrCreate0.id = randomUUID(),
+                        this0_movies_connectOrCreate0.title = $this0_movies_connectOrCreate_param1
+                    MERGE (this0)-[this0_movies_connectOrCreate_this0:ACTED_IN]->(this0_movies_connectOrCreate0)
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this1.screentime = $this0_movies_connectOrCreate_param2
+                        this0_movies_connectOrCreate_this0.screentime = $this0_movies_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN this0
@@ -317,13 +317,13 @@ describe("Create or Connect", () => {
                 WITH this0
                 CALL {
                     WITH this0
-                    MERGE (this0_movies_connectOrCreate_this0:\`Movie\` { id: $this0_movies_connectOrCreate_param0 })
+                    MERGE (this0_movies_connectOrCreate0:\`Movie\` { id: $this0_movies_connectOrCreate_param0 })
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this0.createdAt = datetime(),
-                        this0_movies_connectOrCreate_this0.title = $this0_movies_connectOrCreate_param1
-                    MERGE (this0)-[this0_movies_connectOrCreate_this1:ACTED_IN]->(this0_movies_connectOrCreate_this0)
+                        this0_movies_connectOrCreate0.createdAt = datetime(),
+                        this0_movies_connectOrCreate0.title = $this0_movies_connectOrCreate_param1
+                    MERGE (this0)-[this0_movies_connectOrCreate_this0:ACTED_IN]->(this0_movies_connectOrCreate0)
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this1.screentime = $this0_movies_connectOrCreate_param2
+                        this0_movies_connectOrCreate_this0.screentime = $this0_movies_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN this0
@@ -380,14 +380,14 @@ describe("Create or Connect", () => {
                 WITH this
                 CALL {
                     WITH this
-                    MERGE (this_movies0_connectOrCreate_this0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
+                    MERGE (this_movies0_connectOrCreate0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this0.createdAt = datetime(),
-                        this_movies0_connectOrCreate_this0.id = randomUUID(),
-                        this_movies0_connectOrCreate_this0.title = $this_movies0_connectOrCreate_param1
-                    MERGE (this)-[this_movies0_connectOrCreate_this1:ACTED_IN]->(this_movies0_connectOrCreate_this0)
+                        this_movies0_connectOrCreate0.createdAt = datetime(),
+                        this_movies0_connectOrCreate0.id = randomUUID(),
+                        this_movies0_connectOrCreate0.title = $this_movies0_connectOrCreate_param1
+                    MERGE (this)-[this_movies0_connectOrCreate_this0:ACTED_IN]->(this_movies0_connectOrCreate0)
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this1.screentime = $this_movies0_connectOrCreate_param2
+                        this_movies0_connectOrCreate_this0.screentime = $this_movies0_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN collect(DISTINCT this { .name }) AS data"
@@ -442,13 +442,13 @@ describe("Create or Connect", () => {
                 WITH this
                 CALL {
                     WITH this
-                    MERGE (this_movies0_connectOrCreate_this0:\`Movie\` { id: $this_movies0_connectOrCreate_param0 })
+                    MERGE (this_movies0_connectOrCreate0:\`Movie\` { id: $this_movies0_connectOrCreate_param0 })
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this0.createdAt = datetime(),
-                        this_movies0_connectOrCreate_this0.title = $this_movies0_connectOrCreate_param1
-                    MERGE (this)-[this_movies0_connectOrCreate_this1:ACTED_IN]->(this_movies0_connectOrCreate_this0)
+                        this_movies0_connectOrCreate0.createdAt = datetime(),
+                        this_movies0_connectOrCreate0.title = $this_movies0_connectOrCreate_param1
+                    MERGE (this)-[this_movies0_connectOrCreate_this0:ACTED_IN]->(this_movies0_connectOrCreate0)
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this1.screentime = $this_movies0_connectOrCreate_param2
+                        this_movies0_connectOrCreate_this0.screentime = $this_movies0_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN collect(DISTINCT this { .name }) AS data"
@@ -541,14 +541,14 @@ describe("Create or Connect", () => {
                 WITH this0
                 CALL {
                     WITH this0
-                    MERGE (this0_movies_connectOrCreate_this0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
+                    MERGE (this0_movies_connectOrCreate0:\`Movie\` { title: $this0_movies_connectOrCreate_param0 })
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this0.title = $this0_movies_connectOrCreate_param1
-                    MERGE (this0)-[this0_movies_connectOrCreate_this1:ACTED_IN]->(this0_movies_connectOrCreate_this0)
+                        this0_movies_connectOrCreate0.title = $this0_movies_connectOrCreate_param1
+                    MERGE (this0)-[this0_movies_connectOrCreate_this0:ACTED_IN]->(this0_movies_connectOrCreate0)
                     ON CREATE SET
-                        this0_movies_connectOrCreate_this1.createdAt = datetime(),
-                        this0_movies_connectOrCreate_this1.id = randomUUID(),
-                        this0_movies_connectOrCreate_this1.screentime = $this0_movies_connectOrCreate_param2
+                        this0_movies_connectOrCreate_this0.createdAt = datetime(),
+                        this0_movies_connectOrCreate_this0.id = randomUUID(),
+                        this0_movies_connectOrCreate_this0.screentime = $this0_movies_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN this0
@@ -605,14 +605,14 @@ describe("Create or Connect", () => {
                 WITH this
                 CALL {
                     WITH this
-                    MERGE (this_movies0_connectOrCreate_this0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
+                    MERGE (this_movies0_connectOrCreate0:\`Movie\` { title: $this_movies0_connectOrCreate_param0 })
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this0.title = $this_movies0_connectOrCreate_param1
-                    MERGE (this)-[this_movies0_connectOrCreate_this1:ACTED_IN]->(this_movies0_connectOrCreate_this0)
+                        this_movies0_connectOrCreate0.title = $this_movies0_connectOrCreate_param1
+                    MERGE (this)-[this_movies0_connectOrCreate_this0:ACTED_IN]->(this_movies0_connectOrCreate0)
                     ON CREATE SET
-                        this_movies0_connectOrCreate_this1.createdAt = datetime(),
-                        this_movies0_connectOrCreate_this1.id = randomUUID(),
-                        this_movies0_connectOrCreate_this1.screentime = $this_movies0_connectOrCreate_param2
+                        this_movies0_connectOrCreate_this0.createdAt = datetime(),
+                        this_movies0_connectOrCreate_this0.id = randomUUID(),
+                        this_movies0_connectOrCreate_this0.screentime = $this_movies0_connectOrCreate_param2
                     RETURN COUNT(*) AS _
                 }
                 RETURN collect(DISTINCT this { .name }) AS data"