+++ packages/graphql/tests/tck/issues/1115.test.ts
@@ -79,21 +79,23 @@ describe("https://github.com/neo4j/graphql/issues/1115", () => {
             WITH this
             CALL {
                 WITH this
-                CALL apoc.util.validate(NOT (any(auth_var1 IN [\\"upstream\\"] WHERE any(auth_var0 IN $auth.roles WHERE auth_var0 = auth_var1))), \\"@neo4j/graphql/FORBIDDEN\\", [0])
-                MERGE (this_connectOrCreate_children_this0:\`Child\` { tcId: $this_connectOrCreate_children_param0 })
+                MERGE (this_connectOrCreate_children0:\`Child\` { tcId: $this_connectOrCreate_children_param0 })
                 ON CREATE SET
-                    this_connectOrCreate_children_this0.tcId = $this_connectOrCreate_children_param1
-                MERGE (this_connectOrCreate_children_this0)-[this_connectOrCreate_children_this1:HAS]->(this)
+                    this_connectOrCreate_children0.tcId = $this_connectOrCreate_children_param1
+                MERGE (this_connectOrCreate_children0)-[this_connectOrCreate_children_this0:HAS]->(this)
+                WITH *
+                CALL apoc.util.validate(NOT (any(auth_var1 IN [\\"upstream\\"] WHERE any(auth_var0 IN $auth.roles WHERE auth_var0 = auth_var1))), \\"@neo4j/graphql/FORBIDDEN\\", [0])
                 RETURN COUNT(*) AS _
             }
             WITH this
             CALL {
                 WITH this
-                CALL apoc.util.validate(NOT (any(auth_var1 IN [\\"upstream\\"] WHERE any(auth_var0 IN $auth.roles WHERE auth_var0 = auth_var1))), \\"@neo4j/graphql/FORBIDDEN\\", [0])
-                MERGE (this_connectOrCreate_children_this2:\`Child\` { tcId: $this_connectOrCreate_children_param2 })
+                MERGE (this_connectOrCreate_children1:\`Child\` { tcId: $this_connectOrCreate_children_param2 })
                 ON CREATE SET
-                    this_connectOrCreate_children_this2.tcId = $this_connectOrCreate_children_param3
-                MERGE (this_connectOrCreate_children_this2)-[this_connectOrCreate_children_this3:HAS]->(this)
+                    this_connectOrCreate_children1.tcId = $this_connectOrCreate_children_param3
+                MERGE (this_connectOrCreate_children1)-[this_connectOrCreate_children_this1:HAS]->(this)
+                WITH *
+                CALL apoc.util.validate(NOT (any(auth_var1 IN [\\"upstream\\"] WHERE any(auth_var0 IN $auth.roles WHERE auth_var0 = auth_var1))), \\"@neo4j/graphql/FORBIDDEN\\", [0])
                 RETURN COUNT(*) AS _
             }
             WITH *